Use SDL_GetTicks64 instead of SDL_GetTicks

+++ Snake/SIntroState.h
@@ -40,7 +40,7 @@ private:
     SDL_Surface* title;
     SDL_Surface* fader;
     int alpha;
-    int startticks;
+    Uint64 startticks;
 };
 
 #endif
